[2.2][bug-fix] - cleaned up gender value set of demographic
